% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartbeat.R
\name{heartbeat_send_signal}
\alias{heartbeat_send_signal}
\title{Send a signal}
\usage{
heartbeat_send_signal(con, key, signal)
}
\arguments{
\item{con}{A hiredis object}

\item{key}{The heartbeat key}

\item{signal}{A signal to send (e.g. `tools::SIGINT` or
`tools::SIGKILL`)}
}
\description{
Sends a signal to a heartbeat process that is using key `key`
}
\examples{
if (redux::redis_available()) {
  rand_str <- function() {
    paste(sample(letters, 20, TRUE), collapse = "")
  }
  # Suppose we have a process that exposes a heartbeat running on
  # this key:
  key <- sprintf("redux:heartbeat:test:\%s", rand_str())

  # We can send it an interrupt over redis using:
  con <- redux::hiredis()
  redux::heartbeat_send_signal(con, key, tools::SIGINT)
}
}
