% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{redis}
\alias{redis}
\title{Redis commands object}
\format{
An object of class \code{redis_commands} of length 199.
}
\usage{
redis
}
\description{
Primarily used for pipelining, the \code{redis} object produces
commands the same way that the main \code{\link[=redis_api]{redis_api()}}
objects do.  If passed in as arguments to the \code{pipeline}
method (where supported) these commands will then be pipelined.
}
\examples{
# This object creates commands in the format expected by the
# lower-level redis connection object:
redis$PING()

# For example to send two PING commands in a single transmission:
if (redux::redis_available()) {
  r <- redux::hiredis()
  r$pipeline(
    redux::redis$PING(),
    redux::redis$PING())
}
}
\keyword{datasets}
