% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis_api.R
\name{redis_api}
\alias{redis_api}
\title{Create a Redis API object}
\usage{
redis_api(x, version = NULL)
}
\arguments{
\item{x}{An object that defines at least the function
\code{command} capable of processing commands in the appropriate
form.}

\item{version}{Version of the Redis API to generate.  If given as
a numeric version (or something that can be coerced into one.
If given as \code{TRUE}, then we query the Redis server for its
version and generate only commands supported by the server.}
}
\description{
Create a Redis API object.  This function is designed to be used
from other packages, and not designed to be used directly by
users.
}
