% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{redis_scripts}
\alias{redis_scripts}
\title{Load Lua scripts into Redis}
\usage{
redis_scripts(con, ..., scripts = list(...))
}
\arguments{
\item{con}{A Redis connection}

\item{...}{A number of scripts}

\item{scripts}{Alternatively, a list of scripts}
}
\description{
Load Lua scripts into Redis, providing a convenience function to
call them with.  Using this function means that scripts will be
available to use via \code{EVALSHA}, and will be preloaded on the Redis
server.  Scripts are then accessed by \emph{name} rather than by
content or SHA.  See \code{vignette("redux")} for details and an
example.
}
