% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis_api.R
\docType{data}
\name{redis}
\alias{redis}
\title{Redis commands object}
\format{An object of class \code{redis_commands} of length 188.}
\usage{
redis
}
\description{
Primarily used for pipeling, the \code{redis} object produces
commands the same way that the main \code{\link{redis_api}}
objects do.  If passed in as arguments to the \code{pipeline}
method (where supported) these commands will then be pipelined.
See the \code{redux} package for an example.
}
\examples{
redis$PING()
}
\keyword{datasets}

