% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/connection.R
\name{redis_connection}
\alias{redis_connection}
\title{Create a Redis connection}
\usage{
redis_connection(config)
}
\arguments{
\item{config}{Configuration parameters as generated by
\code{\link[RedisAPI]{redis_config}}.}
}
\description{
Create a Redis connection.  This function is designed to be used
in other packages, and not directly by end-users.  However, it is
possible and safe to use.  See the \code{RedisAPI} package for a
complete example.
}
\details{
This function creates a list of functions, appropriately bound to
a pointer to a Redis connection.  This is designed for package
authors to use so without having to ever deal with the actual
pointer itself (which cannot be directly manipulated from R
anyway).

The returned list has elements, all of which are functions:

\describe{
\item{\code{config()}}{The configuration information}

\item{\code{reconnect()}}{Attempt reconnection of a connection
that has been closed, through serialisation/deserialiation or
through loss of internet connection.}

\item{command(cmd)}{Run a Redis command.  The format of this
command will be documented elsewhere.}

\item{pipeline(cmds)}{Run a pipeline of Redis commands.}

\item{subscribe(channel, callback, envir)}{this will change...}

}
}

