% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis_tools.R
\name{redis_info}
\alias{parse_info}
\alias{redis_info}
\alias{redis_version}
\title{Parse Redis INFO}
\usage{
redis_info(con)

parse_info(x)

redis_version(con)
}
\arguments{
\item{con}{A Redis connection}

\item{x}{character string}
}
\description{
Parse and return Redis \code{INFO} data.
}

