% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis_tools.R
\name{redis_multi}
\alias{redis_multi}
\title{Helper for Redis MULTI}
\usage{
redis_multi(con, expr)
}
\arguments{
\item{con}{A Redis connection object}

\item{expr}{An expression to evaluate}
}
\description{
Helper to evaluate a Redis \code{MULTI} statement.  If an error
occurs then, \code{DISCARD} is called and the transaction is
cancelled.  Otherwise \code{EXEC} is called and the transaction is
processed.
}

