% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis_tools.R
\name{redis_time}
\alias{format_redis_time}
\alias{redis_time}
\alias{redis_time_to_r}
\title{Get time from Redis}
\usage{
redis_time(con)

format_redis_time(x)

redis_time_to_r(x)
}
\arguments{
\item{con}{A Redis connection object}

\item{x}{a list as returned by \code{TIME}}
}
\description{
Get time from Redis and format as a string.
}

